Feature/refactor services directory (#123)
* first commit of input service, tests still broken

* fix question component tests

* remove comment and fix input value for touch events

* putting back keydown handler on check component

* improvements

* record backspace from virtual keyboard

* 14516 tests

* remove failing test

* register enter as input while onSubmit

* lint fixes

* Integrate answer service on submit and timeout calls

* 14928 Resolved tests

* define answer object before submitting to service

* Refactor services to separate directory

* Move services to subfolders
