+++ pupil-spa/src/app/logout/logout.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
-import { UserService } from '../user.service';
+import { UserService } from '../services/user/user.service';
 import { Router } from '@angular/router';
-import { QuestionService } from '../question.service';
+import { QuestionService } from '../services/question/question.service';
 
 @Component({
   selector: 'app-logout',
