+++ pupil-spa/src/app/services/answer/answer.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { StorageService } from './storage.service';
+import { StorageService } from '../storage/storage.service';
 
 export class Answer {
