+++ pupil-spa/src/app/services/audit/audit.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { StorageService } from './storage.service';
+import { StorageService } from '../storage/storage.service';
 import * as AuditTypes from './auditEntry';
 
 @Injectable()
