+++ pupil-spa/src/app/services/question/question.service.spec.ts
@@ -1,8 +1,8 @@
 import { TestBed, inject } from '@angular/core/testing';
 
 import { QuestionService } from './question.service';
-import { StorageService } from './storage.service';
-import * as responseMock from './login.response.mock.json';
+import { StorageService } from '../storage/storage.service';
+import * as responseMock from '../../login.response.mock.json';
 
 describe('QuestionService', () => {
