+++ pupil-spa/src/app/services/question/question.service.ts
@@ -1,8 +1,8 @@
 import { Injectable } from '@angular/core';
 import { Question } from './question.model';
 
-import { StorageService } from './storage.service';
-import { Config } from './config.model';
+import { StorageService } from '../storage/storage.service';
+import { Config } from '../../config.model';
 
 
 @Injectable()
