+++ pupil-spa/src/app/services/user/user.service.spec.ts
@@ -8,8 +8,8 @@ import {
 } from '@angular/http';
 import { MockBackend } from '@angular/http/testing';
 import { UserService } from './user.service';
-import { StorageService } from './storage.service';
-import * as mockLoginResponseBody from './login.userService.response.mock.json';
+import { StorageService } from '../storage/storage.service';
+import * as mockLoginResponseBody from '../../login.userService.response.mock.json';
 
 const shouldNotExecute = () => {
   expect('this code').toBe('not executed');
